+++ packages/hub-nodejs/package.json
@@ -23,6 +23,7 @@
     "@farcaster/core": "0.14.11",
     "@grpc/grpc-js": "~1.8.21",
     "@noble/hashes": "^1.3.0",
+    "@pinojs/json-colorizer": "^4.0.0",
     "neverthrow": "^6.0.0"
   },
   "scripts": {
@@ -34,14 +35,15 @@
     "test": "NODE_OPTIONS=--experimental-vm-modules jest",
     "test:ci": "ENVIRONMENT=test NODE_OPTIONS=--experimental-vm-modules jest --ci --forceExit --coverage",
     "prepublishOnly": "yarn run build",
+    "register-accounts": "tsx src/scripts/register-accounts.ts",
     "test-script": "tsx src/scripts/test.ts",
     "start:chain": "yarn hardhat node --config hardhat.config.cjs"
   },
   "devDependencies": {
     "@faker-js/faker": "^7.6.0",
     "biome-config-custom": "*",
-    "ts-proto": "^1.146.0",
-    "hardhat": "^2.19.2"
+    "hardhat": "^2.19.2",
+    "ts-proto": "^1.146.0"
   },
   "publishConfig": {
     "access": "public"
