+++ packages/hub-nodejs/src/scripts/test.ts
@@ -1,194 +1,65 @@
-import * as ed from "@noble/ed25519";
 import {
-  ID_GATEWAY_ADDRESS,
-  ID_REGISTRY_ADDRESS,
-  ViemLocalEip712Signer,
-  idGatewayABI,
-  idRegistryABI,
-  NobleEd25519Signer,
-  BUNDLER_ADDRESS,
-  bundlerABI,
-  KEY_GATEWAY_ADDRESS,
-  keyGatewayABI,
   FarcasterNetwork,
+  getInsecureHubRpcClient,
+  makeCastAdd,
+  NobleEd25519Signer,
 } from "@farcaster/hub-nodejs";
-import { bytesToHex, createPublicClient, createWalletClient, http, parseEther, toBytes } from "viem";
-import { privateKeyToAccount, generatePrivateKey } from "viem/accounts";
-import { optimism } from "viem/chains";
+import { hexToBytes } from "@noble/hashes/utils";
 
+/**
+ * Populate the following constants with your own values
+ */
+
+// Signer private key registered with the Hub (see hello-world example)
+const SIGNER = "0xf7baa925cbadc673ae3eb6fa33fe61f4134acfb003ce33e6df6d19fe3ef34684";
+// Fid owned by the custody address
+const FID = 586432; // <REQUIRED>
+
+// Testnet Configuration
 const HUB_URL = "localhost:2283"; // URL + Port of the Hub
 const NETWORK = FarcasterNetwork.DEVNET; // Network of the Hub
 
-// 1. Set up clients and account keys
-
-const hardhatAccount = privateKeyToAccount("0xac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80");
-
-const APP_PRIVATE_KEY = generatePrivateKey();
-const APP_PUBLIC_KEY = privateKeyToAccount(APP_PRIVATE_KEY);
-const ALICE_PRIVATE_KEY = generatePrivateKey();
-const ALICE_PUBLIC_KEY = privateKeyToAccount(ALICE_PRIVATE_KEY);
-const OP_PROVIDER_URL = "http://localhost:8545";
-
-const publicClient = createPublicClient({
-  chain: optimism,
-  transport: http(OP_PROVIDER_URL),
-});
-
-const walletClient = createWalletClient({
-  chain: optimism,
-  transport: http(OP_PROVIDER_URL),
-});
-
-const app = privateKeyToAccount(APP_PRIVATE_KEY);
-const appAccountKey = new ViemLocalEip712Signer(app);
-
-const alice = privateKeyToAccount(ALICE_PRIVATE_KEY);
-const aliceAccountKey = new ViemLocalEip712Signer(alice);
-
-const deadline = BigInt(Math.floor(Date.now() / 1000) + 3600); // set the signatures' deadline to 1 hour from now
-
-const WARPCAST_RECOVERY_PROXY = "0x00000000FcB080a4D6c39a9354dA9EB9bC104cd7";
-
-async function sendEth(address: `0x${string}`, value: string) {
-  const sendTxHash = await walletClient.sendTransaction({
-    account: hardhatAccount.address,
-    to: address,
-    value: parseEther(value),
-  });
-  console.log("Sent eth", address, parseEther(value));
-}
-
 (async () => {
-  await sendEth(APP_PUBLIC_KEY.address, "100");
-  await sendEth(ALICE_PUBLIC_KEY.address, "100");
+  // Set up the signer
+  const privateKeyBytes = hexToBytes(SIGNER.slice(2));
+  const ed25519Signer = new NobleEd25519Signer(privateKeyBytes);
+  // const signerPublicKey = (await ed25519Signer.getSignerKey())._unsafeUnwrap();
 
-  // 2. Register an app fid
+  const dataOptions = {
+    fid: FID,
+    network: NETWORK,
+  };
 
-  const price = await publicClient.readContract({
-    address: ID_GATEWAY_ADDRESS,
-    abi: idGatewayABI,
-    functionName: "price",
-    args: [0n],
-  });
+  // If your client does not use SSL.
+  const client = getInsecureHubRpcClient(HUB_URL);
 
-  const { request } = await publicClient.simulateContract({
-    account: app,
-    address: ID_GATEWAY_ADDRESS,
-    abi: idGatewayABI,
-    functionName: "register",
-    args: [WARPCAST_RECOVERY_PROXY, 0n],
-    value: price,
-  });
-  await walletClient.writeContract(request);
+  const f = await client.getFids({});
+  console.log(f);
 
-  const APP_FID = await publicClient.readContract({
-    address: ID_REGISTRY_ADDRESS,
-    abi: idRegistryABI,
-    functionName: "idOf",
-    args: [app.address],
-  });
+  /**
+   * Example 1: A cast with no mentions
+   *
+   * "This is a cast with no mentions"
+   */
 
-  // 3. Collect a Register signature from the user.
+  const castAdd = await makeCastAdd(
+    {
+      text: "This is a cast with no mentions",
+      embeds: [],
+      embedsDeprecated: [],
+      mentions: [],
+      mentionsPositions: [],
+    },
+    dataOptions,
+    ed25519Signer,
+  );
+  // castResults.push(cast);
 
-  let nonce = await publicClient.readContract({
-    address: KEY_GATEWAY_ADDRESS,
-    abi: keyGatewayABI,
-    functionName: "nonces",
-    args: [alice.address],
-  });
+  const cast = await client.submitMessage(castAdd._unsafeUnwrap());
 
-  const registerSignatureResult = await aliceAccountKey.signRegister({
-    to: alice.address as `0x${string}`,
-    recovery: WARPCAST_RECOVERY_PROXY,
-    nonce,
-    deadline,
-  });
+  const y = await client.getCastsByFid({ fid: FID });
 
-  let registerSignature;
-  if (registerSignatureResult.isOk()) {
-    registerSignature = registerSignatureResult.value;
-  } else {
-    throw new Error("Failed to generate register signature");
-  }
+  console.log(JSON.stringify(y, null, 2));
 
-  // Create an Ed25519 account keypair for Alice and get the public key.
-  const privateKeyBytes = ed.utils.randomPrivateKey();
-  const accountKey = new NobleEd25519Signer(privateKeyBytes);
-  let accountPubKey = new Uint8Array();
-  const accountKeyResult = await accountKey.getSignerKey(); // why doesn't aliceAccountKey.getSignerKey(); work???
-
-  // 5. Use your app account to create a Signed Key Request
-  if (accountKeyResult.isOk()) {
-    accountPubKey = accountKeyResult.value;
-
-    const signedKeyRequestMetadata = await appAccountKey.getSignedKeyRequestMetadata({
-      requestFid: APP_FID,
-      key: accountPubKey,
-      deadline,
-    });
-
-    // 6. Collect an Add signature from the user.
-    if (signedKeyRequestMetadata.isOk()) {
-      const metadata = bytesToHex(signedKeyRequestMetadata.value);
-
-      nonce = await publicClient.readContract({
-        address: KEY_GATEWAY_ADDRESS,
-        abi: keyGatewayABI,
-        functionName: "nonces",
-        args: [alice.address],
-      });
-
-      const addSignatureResult = await aliceAccountKey.signAdd({
-        owner: alice.address as `0x${string}`,
-        keyType: 1,
-        key: accountPubKey,
-        metadataType: 1,
-        metadata,
-        nonce,
-        deadline,
-      });
-
-      // 7. Call the Bundler contract to register onchain.
-      if (addSignatureResult.isOk()) {
-        const addSignature = addSignatureResult.value;
-
-        console.log(addSignature);
-
-        const price = await publicClient.readContract({
-          address: BUNDLER_ADDRESS,
-          abi: bundlerABI,
-          functionName: "price",
-          args: [0n],
-        });
-
-        const { request } = await publicClient.simulateContract({
-          account: app,
-          address: BUNDLER_ADDRESS,
-          abi: bundlerABI,
-          functionName: "register",
-          args: [
-            {
-              to: alice.address,
-              recovery: WARPCAST_RECOVERY_PROXY,
-              sig: bytesToHex(registerSignature),
-              deadline,
-            },
-            [
-              {
-                keyType: 1,
-                key: bytesToHex(accountPubKey),
-                metadataType: 1,
-                metadata: metadata,
-                sig: bytesToHex(addSignature),
-                deadline,
-              },
-            ],
-            0n,
-          ],
-          value: price,
-        });
-        await walletClient.writeContract(request);
-      }
-    }
-  }
+  client.close();
 })();
